requirements.txt
Django==2.1.5
django-appconf==1.0.2
django-compressor==2.2
django-filebrowser-no-grappelli==3.7.7
django-geoposition==0.3.0
django-jquery==3.1.0
django-libsass==0.7
django-ordered-model==3.1.1
django-tinymce==2.8.0
libsass==0.17.0
Pillow==5.4.1
psycopg2==2.8.4
pytz==2018.9
rcssmin==1.0.6
rjsmin==1.0.12
six==1.12.0
